+++ test/functional/mac2/driver_test.rb
@@ -20,6 +20,9 @@ class AppiumLibCoreTest
     def setup
       @core = ::Appium::Core.for(Caps.mac2)
       @driver = @core.start_driver
+
+      require 'pry'
+      binding.pry
     end
 
     def teardown