+++ CHANGELOG.md
@@ -5,8 +5,9 @@ Read `release_notes.md` for commit level details.
 ## [Unreleased]
 
 ### Enhancements
-- Add `Element#screenshot_as` and `Element#save_screenshot` in Element module
-    - They are same as `Driver#element_screenshot_as` and `Driver#save_element_screenshot`
+- Add  `Element#screenshot`, `Element#screenshot_as` and `Element#save_screenshot` in Element module
+    - `Element#screenshot_as` and `Element#save_screenshot` are same as `Driver#element_screenshot_as` and `Driver#save_element_screenshot`
+    - `Element#screenshot` is same as `Element#screenshot_as(:base64)`
 
 ### Bug fixes